checks if an element in a mega menu is visible
This is a script that checks if an a text "shoes" is visible in the mega menu of a shop. If it is not, then it takes a screenshot and send information with the time the problem was detected.
This script can be run every few minutes, e.g. by cronjob, as a way of monitoring the problem.